Remove V1, relock, fix imports, adjust conftest
