+++ tests/test_task.py
@@ -1,19 +1,16 @@
-from typing import Dict, List, Tuple
-
 import pytest
 from argo.workflows.client import V1alpha1Arguments, V1alpha1Inputs, V1Toleration
 from pydantic import BaseModel, ValidationError
 
-from hera.v1.artifact import InputArtifact, OutputArtifact
-from hera.v1.empty_dir_volume import EmptyDirVolume
-from hera.v1.existing_volume import ExistingVolume
-from hera.v1.input import InputFrom
-from hera.v1.operator import Operator
-from hera.v1.resources import Resources
-from hera.v1.retry import Retry
-from hera.v1.task import Task
-from hera.v1.toleration import GPUToleration
-from hera.v1.volume import Volume
+from hera.empty_dir_volume import EmptyDirVolume
+from hera.existing_volume import ExistingVolume
+from hera.input import InputFrom
+from hera.operator import Operator
+from hera.resources import Resources
+from hera.retry import Retry
+from hera.task import Task
+from hera.toleration import GPUToleration
+from hera.volume import Volume
 
 
 def test_next_and_shifting_set_correct_dependencies(no_op):
